Pdf_Analyser/requirements.txt
phidata
python-dotenv
duckduckgo-search
fastapi
groq
